+++ packages/transaction-controller/src/utils/prepare.ts
@@ -4,8 +4,9 @@ import type { TypedTransaction, TypedTxData } from '@ethereumjs/tx';
 import { TransactionFactory } from '@ethereumjs/tx';
 import { bytesToHex } from '@metamask/utils';
 import type { Hex } from '@metamask/utils';
+import { cloneDeep } from 'lodash';
 
-import type { TransactionParams } from '../types';
+import type { AuthorizationList, TransactionParams } from '../types';
 
 export const HARDFORK = Hardfork.Prague;
 
@@ -20,8 +21,10 @@ export function prepareTransaction(
   chainId: Hex,
   txParams: TransactionParams,
 ): TypedTransaction {
+  const normalizedData = normalizeParams(txParams);
+
   // Does not allow `gasPrice` on type 4 transactions.
-  const data = txParams as TypedTxData;
+  const data = normalizedData as TypedTxData;
 
   return TransactionFactory.fromTxData(data, {
     freeze: false,
@@ -54,4 +57,52 @@ function getCommonConfiguration(chainId: Hex): Common {
   return Common.custom(customChainParams, {
     eips: [7702],
   });
+}
+
+/**
+ * Normalize the transaction parameters for compatibility with `ethereumjs/tx`.
+ *
+ * @param params - The transaction parameters to normalize.
+ * @returns The normalized transaction parameters.
+ */
+function normalizeParams(params: TransactionParams): TransactionParams {
+  const newParams = cloneDeep(params);
+  normalizeAuthorizationList(newParams.authorizationList);
+  return newParams;
+}
+
+/**
+ * Normalize the authorization list for `ethereumjs/tx` compatibility.
+ *
+ * @param authorizationList - The list of authorizations to normalize.
+ */
+function normalizeAuthorizationList(authorizationList?: AuthorizationList) {
+  if (!authorizationList) {
+    return;
+  }
+
+  for (const authorization of authorizationList) {
+    authorization.nonce = removeLeadingZeroes(authorization.nonce);
+    authorization.r = removeLeadingZeroes(authorization.r);
+    authorization.s = removeLeadingZeroes(authorization.s);
+    authorization.yParity = removeLeadingZeroes(authorization.yParity);
+  }
+}
+
+/**
+ * Remove leading zeroes from a hexadecimal string.
+ *
+ * @param value - The hexadecimal string to process.
+ * @returns The processed hexadecimal string.
+ */
+function removeLeadingZeroes(value: Hex | undefined): Hex | undefined {
+  if (!value) {
+    return value;
+  }
+
+  if (value === '0x0') {
+    return '0x';
+  }
+
+  return (value.replace?.(/^0x(00)+/u, '0x') as Hex) ?? value;
 }